+++ src/pages/notifications/notifications-page/notifications.component.ts
@@ -1,5 +1,7 @@
 import { Component } from '@angular/core';
 import { NavParams, NavController } from 'ionic-angular';
+import { Subject, Subscription } from 'rxjs';
+import { FirebaseService } from '../../../environments/firebase/firebase-service';
 
 @Component({
     templateUrl: 'notifications.html',
@@ -8,14 +10,34 @@ import { NavParams, NavController } from 'ionic-angular';
 
 export class NotificationsPage {
 
-    notificationsObj;
+    notifications;
+    notificationsCount;
+    count$ = new Subject();
 
-    constructor(private nav: NavController, public navParams: NavParams, ) {
-        this.notificationsObj = navParams.get('myNotifications');
+    constructor(private nav: NavController,
+        public navParams: NavParams,
+        public firebaseService: FirebaseService) {
+
+        var userId = localStorage.getItem('securityToken');
+
+        if (userId !== null) {
+            this.notifications = [];
+            this.firebaseService.getNotifications(userId).subscribe(notifications => {
+                this.notifications = notifications.filter(notification => {
+                    return notification.isread === false;
+                });
+                console.log('this.notifications', this.notifications);
+                this.count$.next(this.notifications.length);
+                this.notificationsCount = this.count$;
+            });
+        }
+    }
+
+    getCount() {
+        return this.count$;
     }
 
     back() {
         this.nav.pop();
     }
-
 }
